Move sqlite to separate library
