+++ src/DBManagment/imp/ConnectionKeeper.cpp
@@ -1,5 +1,5 @@
 #include "../ConnectionKeeper.h"
-#include "../sqlite3.h"
+#include "sqlite3.h"
 #include <stdio.h>
 sqlite3* db;
