+++ src/Repositories/imp/HorseRepository.cpp
@@ -1,6 +1,6 @@
 #include "../HorseRepository.h"
 #include <vector>
-#include "../../DBManagment/sqlite3.h"
+#include "sqlite3.h"
 #include "../../DBManagment/ConnectionKeeper.h"
 #include <string.h>
