+++ src/Repositories/imp/JockeyRepository.cpp
@@ -1,5 +1,5 @@
 #include "../JockeyRepository.h"
-#include "../../DBManagment/sqlite3.h"
+#include "sqlite3.h"
 #include "../../DBManagment/ConnectionKeeper.h"
 #include <string.h>
