+++ src/Repositories/imp/OwnerRepository.cpp
@@ -1,5 +1,5 @@
 #include "../OwnerRepository.h"
-#include "../../DBManagment/sqlite3.h"
+#include "sqlite3.h"
 #include "../../DBManagment/ConnectionKeeper.h"
 #include <string>
 #include <string.h>
