+++ src/Repositories/imp/PrizeRepository.cpp
@@ -1,6 +1,6 @@
 #include "../PrizeRepository.h"
 #include "../RaceRecordRepository.h"
-#include "../../DBManagment/sqlite3.h"
+#include "sqlite3.h"
 #include "../../DBManagment/ConnectionKeeper.h"
 #include "../../Models/RaceRecord.h"
 #include <algorithm>
