+++ src/Repositories/imp/RaceRecordRepository.cpp
@@ -1,5 +1,5 @@
 #include "../RaceRecordRepository.h"
-#include "../../DBManagment/sqlite3.h"
+#include "sqlite3.h"
 #include "../../DBManagment/ConnectionKeeper.h"
 #include <string.h>
