+++ src/Repositories/imp/RaceRepository.cpp
@@ -1,5 +1,5 @@
 #include "../RaceRepository.h"
-#include "../../DBManagment/sqlite3.h"
+#include "sqlite3.h"
 #include "../../DBManagment/ConnectionKeeper.h"
 #include <string>
 #include <string.h>
