+++ test/TestFixture.cpp
@@ -3,7 +3,7 @@
 #include <gtest/gtest.h>
 #include "Repositories/HorseRepository.h"
 #include "DBManagment/ConnectionKeeper.h"
-#include "DBManagment/sqlite3.h"
+#include "sqlite3.h"
 #include <filesystem>
 
 static const std::string db_path = "../../test/test.db";
